fix(mastracode): address startup review feedback
Preserve pinned task state unless bounded history contains a task snapshot, handle deferred browser state errors, colocate LibSQL DB migration coverage, and document Harness.setBrowser usage.
